requirements.txt
python-dotenv
typesense
